refactor: move 7 kube leaf packages from MCO to root pkg/kube
Relocate the following packages from mongodb-community-operator/pkg/kube/
to pkg/kube/ at repo root:
- annotations
- configmap
- lifecycle
- pod
- probes
- resourcerequirements
- persistentvolumeclaim

All 7 are pure leaves (zero MCO imports). Behaviour-preserving refactor -
code unchanged, only import paths rewritten.

Foundation step for the mongodb-community-operator decoupling
(spec: docs/design/decouple-mdb-community-investigation.md, cluster 1).
