+++ mongodb-community-operator/pkg/agent/agent_readiness.go
@@ -12,7 +12,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apiErrors "k8s.io/apimachinery/pkg/api/errors"
 
-	"github.com/mongodb/mongodb-kubernetes/mongodb-community-operator/pkg/kube/pod"
+	"github.com/mongodb/mongodb-kubernetes/pkg/kube/pod"
 )
 
 const (
